[feature] 닉네임 입력 화면 설계
