--- domain/src/main/java/pokitmons/pokit/domain/MyClass.kt
@@ -1,3 +1,0 @@
-package pokitmons.pokit.domain
-
-class MyClass
