+++ feature/login/src/main/java/pokitmons/pokit/LoginState.kt
@@ -5,6 +5,6 @@ import pokitmons.pokit.domain.model.auth.SNSLogin
 sealed class LoginState {
         data object Init : LoginState()
         data class Login(val loginResponse: SNSLogin) : LoginState()
-//        data class CheckDuplicateNickname(val checkDuplicateNicknameResponse: CheckDuplicateNickname)
+        data class CheckDuplicateNickname(val checkDuplicateNicknameResponse: CheckDuplicateNickname) : LoginState()
         data object Error : LoginState()
     }
