+++ feature/login/src/main/java/pokitmons/pokit/LoginViewModel.kt
@@ -3,6 +3,10 @@ package pokitmons.pokit
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.coroutines.Job
+import kotlinx.coroutines.delay
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 import org.orbitmvi.orbit.Container
 import org.orbitmvi.orbit.ContainerHost
@@ -17,9 +21,19 @@ class LoginViewModel @Inject constructor(
     private val loginUseCase: SNSLoginUseCase
 ) : ViewModel(), ContainerHost<LoginState, Nothing> {
 
+    private var apiRequestJob: Job? = null
+
     override val container: Container<LoginState, Nothing> = container(
         initialState = LoginState.Init,
     )
+
+    private val _inputNicknameState = MutableStateFlow("")
+    val inputNicknameState: StateFlow<String>
+        get() = _inputNicknameState
+
+    fun inputText(text: String) {
+        _inputNicknameState.value = text
+    }
 
     fun snsLogin(authPlatform: String, idToken: String) = intent {
         viewModelScope.launch {
@@ -37,8 +51,10 @@ class LoginViewModel @Inject constructor(
     }
 
     fun checkDuplicateNickname(nickname: String) {
-        viewModelScope.launch {
-
+        apiRequestJob?.cancel()
+        apiRequestJob = viewModelScope.launch {
+            delay(1.second())
+            // TOOD api 연동
         }
     }
 
@@ -52,4 +68,11 @@ class LoginViewModel @Inject constructor(
 
     var nickname: String = ""
         private set
+
+    // TODO 확장함수 모듈 생성하기
+    companion object {
+        private fun Int.second(): Long {
+            return (this * 1000L)
+        }
+    }
 }
